requirements.txt
ccxt
python-dotenv
pandas
numpy
ta
Flask
google-genai
